Moved functions to ``pipeline_helper.py``
* Moved functions in ``pipeline.py`` outside ``run`` to
  ``pipeline_helper.py``

* Renamed ``TestPipelineMethods`` to ``TestPipelineHelpers``, and
  modified tests accordingly
